Now have working projections and some encouragemnet/tips for saving
